Split test_regex_sync.py at the seam it already had (#352)
One 1306-line module held two test suites that shared, by AST, exactly
one symbol: _SANCTIONED_EXTRAS, which is about the differential rules'
character classes and belongs on the ledger side anyway. Nothing else
crossed in either direction. The module docstring's first three
paragraphs existed to explain why the two lived together, which is a
filename's job.

test_regex_sync.py keeps what its name says: the pipeline's hand copies
of nameparser.config patterns, 160 lines.

test_ledger_guards.py takes the differential half unchanged -- the hand
copies made inside expected_since_*.toml, and the corpus-derived checks
on what each rule may claim.

_differential_fixtures.py holds the handles on tools/differential/ that
both it and test_differential.py need: the ledger glob, the corpus, and
the compare.py path-loader. That loader existed twice; now once.

A pure move, and checked as one: pytest collects 2387 tests in tests/v2
before and after, and every node id is identical once the module name
is stripped. No test added, removed or changed.

Deliberately NOT split further. Spans, vocabularies and claims stay in
one file, because that three-way category axis is the one six review
rounds proved wrong -- every widening they found landed just outside
whichever category the last guard covered, and two findings existed
because no reader could see two guards' coverage side by side. `grep
'^def test_'` on one file answers "what is guarded"; across three files
it does not.

Also repoints nine cross-references that named test_regex_sync.py for
things that moved -- four ledger comments, the corpus builder, the
harness README, and test_differential.py's own note about the loader.
The four references left pointing at test_regex_sync.py are the
pipeline-copy ones, which are still right.
